requirements.txt
attrs==21.2.0
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.5
dependency-injector==4.32.3
factory-boy==3.2.0
Faker==8.6.0
fastapi==0.65.1
fastapi-jwt-auth==0.5.0
greenlet==1.1.0
h11==0.12.0
httptools==0.1.2
idna==2.10
iniconfig==1.1.1
kafka-python==2.0.2
Mako==1.1.4
MarkupSafe==2.0.1
mongoengine==0.23.1
mongomock==3.23.0
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
pydantic==1.8.2
PyJWT==1.7.1
pymongo==3.11.4
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.4
pytest-env==0.6.2
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.1
python-editor==1.0.4
PyYAML==5.4.1
requests==2.25.1
sentinels==1.0.0
six==1.15.0
SQLAlchemy==1.4.27
starlette==0.14.2
text-unidecode==1.3
toml==0.10.2
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.13.4
uvloop==0.16.0
watchgod==0.7
websockets==8.1
